Implement LookupProcessor for enhanced schema processing and integrate it into SchemaProcessorManager
